apps/backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
httpx
sqlalchemy>=2.0.0
asyncpg>=0.29.0
python-multipart
